Poängsystem implementerat
Spelaren får nu poäng efter avslutat spel, varje rätt svar ger 100 poäng, fel svar ger mindre poäng. Metoden för att beräkna poängen kommer antagligen att behöva ändras och förbättras

+++ VaderLeken/vaderleken/ScoreBoardTableViewCell.swift
@@ -11,7 +11,7 @@ import UIKit
 class ScoreBoardTableViewCell: UITableViewCell {
     @IBOutlet weak var ScoreLabel: UILabel!
     
-    func showScore(score: Int){
+    func showScore(score: Double){
         ScoreLabel.text = String(score)
     }
